[SYNTH-12989] Unexport `getTestsToTrigger()`
