+++ src/commands/synthetics/run-tests-lib.ts
@@ -26,13 +26,12 @@ import {updateLTDMultiLocators} from './multilocator'
 import {DefaultReporter, getTunnelReporter} from './reporters/default'
 import {JUnitReporter} from './reporters/junit'
 import {DEFAULT_BATCH_TIMEOUT, DEFAULT_COMMAND_CONFIG} from './run-tests-command'
-import {getTestConfigs, getTestsFromSearchQuery} from './test'
+import {getTestConfigs, getTestsFromSearchQuery, getTestsToTrigger} from './test'
 import {Tunnel} from './tunnel'
 import {getTriggerConfigPublicId, isLocalTriggerConfig} from './utils/internal'
 import {
   getReporter,
   getOrgSettings,
-  getTestsToTrigger,
   InitialSummary,
   renderResults,
   getExitReason,
