+++ src/commands/synthetics/utils/public.ts
@@ -9,8 +9,7 @@ import glob from 'glob'
 
 import {getCommonAppBaseURL} from '../../../helpers/app'
 
-import {APIHelper, formatBackendErrors, getApiHelper} from '../api'
-import {replaceConfigWithTestOverrides} from '../compatibility'
+import {formatBackendErrors, getApiHelper} from '../api'
 import {CiError, CriticalError} from '../errors'
 import {
   APIHelperConfig,
@@ -32,14 +31,11 @@ import {
   TriggerConfig,
   UserConfigOverride,
 } from '../interfaces'
-import {uploadMobileApplicationsAndUpdateOverrideConfigs} from '../mobile'
-import {DEFAULT_POLLING_TIMEOUT, MAX_TESTS_TO_TRIGGER} from '../run-tests-command'
-import {getTestAndOverrideConfig} from '../test'
+import {DEFAULT_POLLING_TIMEOUT} from '../run-tests-command'
 
 import {
   LOCAL_TEST_DEFINITION_PUBLIC_ID_PLACEHOLDER,
   hasDefinedResult,
-  isMobileTestWithOverride,
   getBasePayload,
   isLocalTriggerConfig,
   wait,
@@ -321,83 +317,6 @@ export const getReporter = (reporters: Reporter[]): MainReporter => ({
 
 export const isDeviceIdSet = (result: ServerResult): result is Required<BrowserServerResult> =>
   'device' in result && result.device !== undefined
-
-// XXX: We shouldn't export functions that take an `APIHelper` because the `utils` module is exported while `api` is not.
-export const getTestsToTrigger = async (
-  api: APIHelper,
-  triggerConfigs: TriggerConfig[],
-  reporter: MainReporter,
-  triggerFromSearch?: boolean,
-  failOnMissingTests?: boolean,
-  isTunnelEnabled?: boolean
-) => {
-  const errorMessages: string[] = []
-
-  // TODO SYNTH-12989: Clean up deprecated `config` in favor of `testOverrides`
-  triggerConfigs = triggerConfigs.map((triggerConfig) => ({
-    ...triggerConfig,
-    testOverrides: replaceConfigWithTestOverrides(triggerConfig.config, triggerConfig.testOverrides),
-  }))
-
-  // When too many tests are triggered, if fetched from a search query: simply trim them and show a warning,
-  // otherwise: retrieve them and fail later if still exceeding without skipped/missing tests.
-  if (triggerFromSearch && triggerConfigs.length > MAX_TESTS_TO_TRIGGER) {
-    const testsCount = triggerConfigs.length
-    triggerConfigs.splice(MAX_TESTS_TO_TRIGGER)
-    const maxTests = chalk.bold(MAX_TESTS_TO_TRIGGER)
-    errorMessages.push(
-      chalk.yellow(`The search query returned ${testsCount} tests, only the first ${maxTests} will be triggered.\n`)
-    )
-  }
-
-  const initialSummary = createInitialSummary()
-  const testsAndConfigsOverride = await Promise.all(
-    triggerConfigs.map((triggerConfig) =>
-      getTestAndOverrideConfig(api, triggerConfig, reporter, initialSummary, isTunnelEnabled)
-    )
-  )
-
-  await uploadMobileApplicationsAndUpdateOverrideConfigs(
-    api,
-    triggerConfigs,
-    testsAndConfigsOverride.filter(isMobileTestWithOverride)
-  )
-
-  const overriddenTestsToTrigger: TestPayload[] = []
-  const waitedTests: Test[] = []
-  testsAndConfigsOverride.forEach((item) => {
-    if ('errorMessage' in item) {
-      errorMessages.push(item.errorMessage)
-    }
-
-    if ('overriddenConfig' in item) {
-      overriddenTestsToTrigger.push(item.overriddenConfig)
-    }
-
-    if ('test' in item) {
-      waitedTests.push(item.test)
-    }
-  })
-
-  // Display errors at the end of all tests for better visibility.
-  reporter.initErrors(errorMessages)
-
-  if (failOnMissingTests && initialSummary.testsNotFound.size > 0) {
-    const testsNotFoundListStr = [...initialSummary.testsNotFound].join(', ')
-    throw new CiError('MISSING_TESTS', testsNotFoundListStr)
-  }
-
-  if (!overriddenTestsToTrigger.length) {
-    throw new CiError('NO_TESTS_TO_RUN')
-  } else if (overriddenTestsToTrigger.length > MAX_TESTS_TO_TRIGGER) {
-    throw new CriticalError(
-      'TOO_MANY_TESTS_TO_TRIGGER',
-      `Cannot trigger more than ${MAX_TESTS_TO_TRIGGER} tests (received ${triggerConfigs.length})`
-    )
-  }
-
-  return {tests: waitedTests, overriddenTestsToTrigger, initialSummary}
-}
 
 export const fetchTest = async (publicId: string, config: SyntheticsCIConfig): Promise<Test> => {
   const apiHelper = getApiHelper(config)
